+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 - Re-export `notify`.
 - Implemented `Debug` for `Hotwatch`.
 - Added `Hotwatch::new_with_custom_delay`.
-- Removed dependency on `derive_more` and `parking_lot`.
+- The background thread will now stop once `Hotwatch` is dropped.
 - Updated docs and added an example.
+- Removed dependency on `derive_more` and `parking_lot`.
 
 # Version 0.4.1 (2019-05-31)